+++ src/main/java/com/sparta/ottoon/auth/dto/PostResponseDto.java
@@ -1,7 +1,7 @@
 package com.sparta.ottoon.auth.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.sparta.ottoon.post.entity.Post;
+import com.sparta.ottoon.auth.entity.Post;
 import jakarta.validation.constraints.NotBlank;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
