+++ src/main/java/com/sparta/ottoon/auth/entity/PostStatus.java
@@ -1,4 +1,4 @@
-package com.sparta.ottoon.post.entity;
+package com.sparta.ottoon.auth.entity;
 
 public enum PostStatus {
     NOTICE("공지"),
